requirements.txt
asgiref==3.6.0
dj-database-url==2.0.0
Django==4.1.5
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-widget-tweaks==1.4.12
importlib-metadata==6.0.0
Markdown==3.4.1
Pillow==9.4.0
psycopg2-binary==2.9.6
python-decouple==3.8
sqlparse==0.4.3
typing_extensions==4.6.2
tzdata==2022.7
zipp==3.12.0
